Extract SumUp settings into settings-sumup.ts
Mirrors the settings-wallets/-domains/-email-templates split: the SumUp
credentials handler and connection-test route now live in their own
module. Promotes the shared testRoute factory (auth + JSON) into
settings-helpers.ts so Stripe, Square, and SumUp all reuse it instead of
duplicating the wrapper.
